% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{calc_jaccard_similarity}
\alias{calc_jaccard_similarity}
\title{Calculate Jaccard similarity}
\usage{
calc_jaccard_similarity(string1, string2)
}
\arguments{
\item{string1}{First string to score similarity}

\item{string2}{Second string to score similarity}
}
\value{
Jaccard similarity score
}
\description{
Calculate the Jaccard similarity between two strings. Strings are first
tokenised, stop words are removed and tokens are stemmed.
}
