% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doilinker-package.R
\docType{package}
\name{doilinker-package}
\alias{doilinker}
\alias{doilinker-package}
\title{doilinker: Link Preprints And Publications By DOI}
\description{
Links preprints to publications using the method described in
    Cabanac G, Oikonomidi T, Boutron I. "Day-to-day discovery of
    preprint-publication links". Scientometrics. 2021;1–20. DOI:
    10.1007/s11192-021-03900-7.
}
\author{
\strong{Maintainer}: Luke Zappia \email{luke@lazappi.id.au} (\href{https://orcid.org/0000-0001-7744-8565}{ORCID})

}
\keyword{internal}
