% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\name{get_cr_authors_str}
\alias{get_cr_authors_str}
\title{Get Crossref authors string}
\usage{
get_cr_authors_str(cr_authors)
}
\arguments{
\item{cr_authors}{Authors data.frame from a Crossref query}
}
\value{
Character vector with authors names
}
\description{
Convert the authors table from Crossref into a single string. It has the form
"LastName1, FirstName1; LastName2, FirstName2;...".
}
