% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\name{get_first_orcid}
\alias{get_first_orcid}
\title{Get first author ORCiD}
\usage{
get_first_orcid(cr_authors)
}
\arguments{
\item{cr_authors}{Authors data.frame from a Crossref query}
}
\value{
Character vector with ORCiD or \code{NA} if not found
}
\description{
Return the ORCiD for the first author of a reference.
}
