% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_doi_preprint}
\alias{is_doi_preprint}
\title{Is DOI preprint}
\usage{
is_doi_preprint(doi, container_title)
}
\arguments{
\item{doi}{DOI to check}

\item{container_title}{Whether or not there is a container title (journal
name) associated with the DOI}
}
\value{
Logical whether the DOI is from a preprint
}
\description{
Check whether a DOI is from a preprint server.
}
