% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_result_match}
\alias{is_result_match}
\title{Is result a match}
\usage{
is_result_match(
  query_title,
  result_title,
  query_orcid,
  result_orcid,
  query_authors,
  result_authors
)
}
\arguments{
\item{query_title}{Title of the query reference}

\item{result_title}{Title of the result reference}

\item{query_orcid}{ORCiD of the query first author}

\item{result_orcid}{ORCiD of the result first author}

\item{query_authors}{Authors string for the query}

\item{result_authors}{Authors string for the result}
}
\value{
Logical whether the result is a match or not
}
\description{
Decide if a results from a DOI query is a match or not.
}
\details{
Result is a match if the similarity between titles is sufficiently high or
if the similarity is lower by the first author ORCiDs match or the first
author names match.
}
