% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\name{same_first_author}
\alias{same_first_author}
\title{Same first author}
\usage{
same_first_author(authors_str1, authors_str2)
}
\arguments{
\item{authors_str1}{Character vector containing first authors list}

\item{authors_str2}{Character vector containing second authors list}
}
\value{
Logical whether first authors are the same
}
\description{
Check if two author strings have the same first author.
}
\details{
Last names of the first authors are compared. If these match then first names
are checked and \code{TRUE} is returned if either the whole first name or first
initial matches. If first and last names cannot be separated the whole author
names are compared. Some simplification of characters in names is performed
to improve matches.
}
