% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_doi_links}
\alias{search_doi_links}
\title{Search DOI links}
\usage{
search_doi_links(
  doi,
  preprint = NULL,
  limit = 20,
  filter_matches = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{doi}{DOI to query}

\item{preprint}{Logical. Whether \code{doi} is a preprint (\code{TRUE}) or publication
(\code{FALSE}). If \code{NULL} (default) attempt to determine this from \code{doi}.}

\item{limit}{Number of Crossref results to return. Default is 20, max is
1000.}

\item{filter_matches}{Logical. If \code{TRUE} only return matches rather than all
Crossref results.}

\item{verbose}{Logical. Whether or not to print progress messages.}
}
\value{
tibble of potential links
}
\description{
Search Crossref for potential linked references for a DOI
}
\details{
Based on the method described in "Day-to-day discovery of
preprint–publication links" https://doi.org/10.1007/s11192-021-03900-7 and
code from https://github.com/gcabanac/preprint-publication-linker. Query
filters have been modified to allow reverse linking publications to preprints
as well as preprints to publications.
}
